+++ lib/src/wire/builders/widget_builder.dart
@@ -132,3 +132,5 @@ part 'widget_builders/checkbox_form_field_builder.dart';
 part 'widget_builders/radio_group_form_field_builder.dart';
 
 part 'widget_builders/overlay_builder.dart';
+
+part 'widget_builders/placeholder_builder.dart';